+++ src/de/deserialize.rs
@@ -51,11 +51,47 @@ pub struct SubAck<'a> {
 }
 
 #[derive(Debug)]
+pub struct PubRec<'a> {
+    /// Packet identifier
+    pub packet_id: u16,
+
+    /// The success status of the publication reception.
+    pub reason_code: u8,
+
+    /// Properties associated with the reception result.
+    pub properties: Vec<Property<'a>, 8>,
+}
+
+#[derive(Debug)]
+pub struct PubComp<'a> {
+    /// Packet identifier
+    pub packet_id: u16,
+
+    /// The success status of the publication completion.
+    pub reason_code: u8,
+
+    /// Properties associated with the completion.
+    pub properties: Vec<Property<'a>, 8>,
+}
+
+#[derive(Debug)]
+pub struct Disconnect<'a> {
+    /// The success status of the disconnection.
+    pub reason_code: u8,
+
+    /// Properties associated with the disconnection.
+    pub properties: Vec<Property<'a>, 8>,
+}
+
+#[derive(Debug)]
 pub enum ReceivedPacket<'a> {
     ConnAck(ConnAck<'a>),
     Publish(Pub<'a>),
     PubAck(PubAck<'a>),
     SubAck(SubAck<'a>),
+    PubRec(PubRec<'a>),
+    PubComp(PubComp<'a>),
+    Disconnect(Disconnect<'a>),
     PingResp,
 }
 
@@ -104,6 +140,12 @@ impl<'a> ReceivedPacket<'a> {
                 }
 
                 Ok(ReceivedPacket::PingResp)
+            }
+
+            MessageType::PubRec => Ok(ReceivedPacket::PubRec(parse_pubrec(packet_reader)?)),
+            MessageType::PubComp => Ok(ReceivedPacket::PubComp(parse_pubcomp(packet_reader)?)),
+            MessageType::Disconnect => {
+                Ok(ReceivedPacket::Disconnect(parse_disconnect(packet_reader)?))
             }
 
             _ => Err(Error::UnsupportedPacket),
@@ -186,6 +228,59 @@ fn parse_suback<const T: usize>(p: &PacketReader<T>) -> Result<SubAck, Error> {
 
     Ok(SubAck {
         packet_identifier: id,
+        reason_code,
+        properties,
+    })
+}
+
+fn parse_pubrec<const T: usize>(p: &PacketReader<T>) -> Result<PubRec, Error> {
+    let id = p.read_u16()?;
+
+    // Reason code and properties are both optionally present.
+    if p.remaining_len()? == 0 {
+        return Ok(PubRec {
+            packet_id: id,
+            reason_code: 0,
+            properties: Vec::new(),
+        });
+    }
+
+    let reason_code = p.read_u8()?;
+    let properties = p.read_properties()?;
+
+    Ok(PubRec {
+        packet_id: id,
+        reason_code,
+        properties,
+    })
+}
+
+fn parse_pubcomp<const T: usize>(p: &PacketReader<T>) -> Result<PubComp, Error> {
+    let id = p.read_u16()?;
+
+    // Reason code and properties are both optionally present.
+    if p.remaining_len()? == 0 {
+        return Ok(PubComp {
+            packet_id: id,
+            reason_code: 0,
+            properties: Vec::new(),
+        });
+    }
+
+    let reason_code = p.read_u8()?;
+    let properties = p.read_properties()?;
+
+    Ok(PubComp {
+        packet_id: id,
+        reason_code,
+        properties,
+    })
+}
+
+fn parse_disconnect<const T: usize>(p: &PacketReader<T>) -> Result<Disconnect, Error> {
+    let reason_code = p.read_u8()?;
+    let properties = p.read_properties()?;
+    Ok(Disconnect {
         reason_code,
         properties,
     })
@@ -310,6 +405,90 @@ mod test {
         let ping_req = ReceivedPacket::parse_message(&mut reader).unwrap();
         match ping_req {
             ReceivedPacket::PingResp => {}
+            _ => panic!("Invalid message"),
+        }
+    }
+
+    #[test]
+    fn deserialize_good_pubcomp() {
+        let mut serialized_pubcomp: [u8; 6] = [
+            7 << 4, // PubComp
+            0x04,   // Remaining length
+            0x00,
+            0x05, // Identifier
+            0x16, // Response Code
+            0x00, // Properties length
+        ];
+        let mut reader = PacketReader::<32>::from_serialized(&mut serialized_pubcomp);
+        let pub_comp = ReceivedPacket::parse_message(&mut reader).unwrap();
+        match pub_comp {
+            ReceivedPacket::PubComp(comp) => {
+                assert_eq!(comp.packet_id, 5);
+                assert_eq!(comp.reason_code, 0x16);
+                assert_eq!(comp.properties.len(), 0);
+            }
+            _ => panic!("Invalid message"),
+        }
+    }
+
+    #[test]
+    fn deserialize_short_pubcomp() {
+        let mut serialized_pubcomp: [u8; 4] = [
+            7 << 4, // PubComp
+            0x02,   // Remaining length
+            0x00,
+            0x05, // Identifier
+        ];
+        let mut reader = PacketReader::<32>::from_serialized(&mut serialized_pubcomp);
+        let pub_comp = ReceivedPacket::parse_message(&mut reader).unwrap();
+        match pub_comp {
+            ReceivedPacket::PubComp(comp) => {
+                assert_eq!(comp.packet_id, 5);
+                assert_eq!(comp.reason_code, 0);
+                assert_eq!(comp.properties.len(), 0);
+            }
+            _ => panic!("Invalid message"),
+        }
+    }
+
+    #[test]
+    fn deserialize_good_pubrec() {
+        let mut serialized_pubrec: [u8; 6] = [
+            5 << 4, // PubRec
+            0x04,   // Remaining length
+            0x00,
+            0x05, // Identifier
+            0x10, // Response Code
+            0x00, // Properties length
+        ];
+        let mut reader = PacketReader::<32>::from_serialized(&mut serialized_pubrec);
+        let pub_rec = ReceivedPacket::parse_message(&mut reader).unwrap();
+        match pub_rec {
+            ReceivedPacket::PubRec(rec) => {
+                assert_eq!(rec.packet_id, 5);
+                assert_eq!(rec.reason_code, 0x10);
+                assert_eq!(rec.properties.len(), 0);
+            }
+            _ => panic!("Invalid message"),
+        }
+    }
+
+    #[test]
+    fn deserialize_short_pubrec() {
+        let mut serialized_pubrec: [u8; 4] = [
+            5 << 4, // PubRec
+            0x02,   // Remaining length
+            0x00,
+            0x05, // Identifier
+        ];
+        let mut reader = PacketReader::<32>::from_serialized(&mut serialized_pubrec);
+        let pub_rec = ReceivedPacket::parse_message(&mut reader).unwrap();
+        match pub_rec {
+            ReceivedPacket::PubRec(rec) => {
+                assert_eq!(rec.packet_id, 5);
+                assert_eq!(rec.reason_code, 0);
+                assert_eq!(rec.properties.len(), 0);
+            }
             _ => panic!("Invalid message"),
         }
     }