+++ src/ser/serialize.rs
@@ -156,6 +156,25 @@ pub fn subscribe_message<'a, 'b, 'c>(
     packet.finalize(MessageType::Subscribe, 0b0010)
 }
 
+pub fn pubrel_message<'c, 'a>(
+    dest: &'c mut [u8],
+    packet_id: u16,
+    reason: u8,
+    properties: &[Property<'a>],
+) -> Result<&'c [u8], Error> {
+    let mut packet = ReversedPacketWriter::new(dest);
+
+    // Properties and reason can be omitted if there reason code is zero and there are no
+    // properties.
+    if !(properties.is_empty() && reason == 0) {
+        packet.write_properties(properties)?;
+        packet.write(&[reason])?;
+    }
+
+    packet.write_u16(packet_id)?;
+    packet.finalize(MessageType::PubRel, 0b0010)
+}
+
 #[test]
 pub fn serialize_publish() {
     let good_publish: [u8; 10] = [
@@ -320,4 +339,38 @@ fn serialize_ping_req() {
 
     let mut buffer: [u8; 1024] = [0; 1024];
     assert_eq!(ping_req_message(&mut buffer).unwrap(), good_ping_req);
+}
+
+#[test]
+fn serialize_pubrel() {
+    let good_pubrel: [u8; 6] = [
+        6 << 4 | 0b10, // PubRec
+        0x04,          // Remaining length
+        0x00,
+        0x05, // Identifier
+        0x10, // Response Code
+        0x00, // Properties length
+    ];
+
+    let mut buffer: [u8; 1024] = [0; 1024];
+    assert_eq!(
+        pubrel_message(&mut buffer, 5, 0x10, &[]).unwrap(),
+        good_pubrel
+    );
+}
+
+#[test]
+fn serialize_short_pubrel() {
+    let good_pubrel: [u8; 4] = [
+        6 << 4 | 0b10, // PubRec
+        0x02,          // Remaining length
+        0x00,
+        0x05, // Identifier
+    ];
+
+    let mut buffer: [u8; 1024] = [0; 1024];
+    assert_eq!(
+        pubrel_message(&mut buffer, 5, 0x0, &[]).unwrap(),
+        good_pubrel
+    );
 }